refactored code for the POI marker
